FHS and packaging helping to setup.py and data files
This patch adds support for a --veusz-resource-dir=foo option which
tells veusz to install its data files in that directory, rather than
in the python path. It also adds --disable-install-docs to prevent
installation of the docs (useful if the packaging does this
separately).

For runtime, the patch allows a resource symlink to be put in the
veusz python directory to point to the data file directory, or to
set $VEUSZ_DATAFILES_DIR to override the default. Note that it
doesn't make the symlink itself. The environment variable is useful
when running the self test, but not having the data files in the
usual way.

The examples directory can also be a symlink pointing to the real
place in the resource directory.