Add ReferenceMultiple class and add master color for xy and nonorthpoint.
We add a new type of reference which can link to multiple
settings. This allows a setting to link to the stylesheet and a
different setting. Relative settings a convert to this kind of setting
when creating a widget.

With this new code we add master color settings for the xy and
nonorthpoint widgets. This allows the user to change the line color,
fill color and error bar colors with one click.