Fix logical errors in transparency in image widget.
Fixes bug #191

Veusz switched over to a RGBA output image if there was a transparent
color in the colormap, or transparency data was provided. There was a
bug in checking for transparent colors, which is now fixed. This patch
also now checks the image data to look for non-finite values which will
also cause a switch to transparent mode.